+++ packages/cli/src/platform/auth/show.ts
@@ -1,8 +1,8 @@
 import { Command } from '@prisma/internals'
 import { green } from 'kleur/colors'
 
+import { argOrThrow } from '../_lib/cli/parameters'
 import { messages } from '../_lib/messages'
-import { argOrThrow } from '../_lib/parameters'
 import { requestOrThrow } from '../_lib/pdp'
 import { getTokenOrThrow, platformParameters } from '../_lib/utils'
