+++ packages/cli/src/platform/environment/create.ts
@@ -1,7 +1,7 @@
 import { Command } from '@prisma/internals'
 
+import { argOrThrow, getOptionalParameter, getRequiredParameterOrThrow } from '../_lib/cli/parameters'
 import { messages } from '../_lib/messages'
-import { argOrThrow, getOptionalParameter, getRequiredParameterOrThrow } from '../_lib/parameters'
 import { requestOrThrow } from '../_lib/pdp'
 import { getTokenOrThrow, platformParameters } from '../_lib/utils'
