+++ packages/cli/src/platform/pulse/disable.ts
