+++ packages/cli/src/platform/pulse/enable.ts
